+++ src/fidesapi/routes/util.py
@@ -1,3 +1,6 @@
+import zlib
+from base64 import urlsafe_b64decode as b64d
+from base64 import urlsafe_b64encode as b64e
 from pathlib import Path
 
 from fastapi import APIRouter
@@ -17,3 +20,14 @@ def get_resource_type(router: APIRouter) -> str:
         The router's resource type
     """
     return router.prefix.replace(f"{API_PREFIX}/", "", 1)
+
+
+def obscure_string(plaintext: str) -> str:
+    "obscures a string as a minor security measure"
+
+    return b64e(zlib.compress(plaintext.encode())).decode()
+
+
+def unobscure_string(obscured: str) -> str:
+    "unobscures a string as a minor security measure"
+    return zlib.decompress(b64d(obscured.encode())).decode()